+++ server.js
@@ -457,11 +457,45 @@ app.use('/pl/course_instance/:course_instance_id/instance_question/:instance_que
 app.use('/pl/course/:course_id', require('./middlewares/authzCourse')); // set res.locals.course
 app.use('/pl/course/:course_id', function(req, res, next) {res.locals.urlPrefix = '/pl/course/' + req.params.course_id; next();});
 app.use('/pl/course/:course_id', function(req, res, next) {res.locals.navbarType = 'course'; next();});
+app.use('/pl/course/:course_id', require('./middlewares/selectOpenIssueCount'));
 app.use(/^\/pl\/course\/[0-9]+\/?$/, function(req, res, _next) {res.redirect(res.locals.urlPrefix + '/overview');}); // redirect plain course URL to overview page
 app.use('/pl/course/:course_id/overview', require('./pages/courseOverview/courseOverview'));
+app.use('/pl/course/:course_id/questions', require('./pages/courseQuestions/courseQuestions'));
+app.use('/pl/course/:course_id/question/:question_id', [
+    require('./middlewares/selectAndAuthzCourseQuestion'),
+    require('./pages/shared/assessmentStatDescriptions'),
+    require('./pages/shared/floatFormatters'),
+    require('./pages/courseQuestion/courseQuestion'),
+]);
+app.use('/pl/course/:course_id/issues', require('./pages/courseIssues/courseIssues'));
 app.use('/pl/course/:course_id/loadFromDisk', require('./pages/instructorLoadFromDisk/instructorLoadFromDisk'));
 app.use('/pl/course/:course_id/syncs', require('./pages/courseSyncs/courseSyncs'));
 app.use('/pl/course/:course_id/jobSequence', require('./pages/instructorJobSequence/instructorJobSequence'));
+
+// clientFiles
+// Omitting clientFilesCourseInstance and clientFilesAssessment, as these are only
+// sensible in the context of a course instance
+app.use('/pl/course/:course_id/clientFilesCourse', require('./pages/clientFilesCourse/clientFilesCourse'));
+app.use('/pl/course/:course_id/question/:question_id/clientFilesQuestion', [
+   require('./middlewares/selectAndAuthzCourseQuestion'),
+   require('./pages/clientFilesQuestion/clientFilesQuestion'),
+   ] );
+
+// generatedFiles
+app.use('/pl/course/:course_id/question/:question_id/generatedFilesQuestion', [
+   require('./middlewares/selectAndAuthzCourseQuestion'),
+   require('./pages/courseGeneratedFilesQuestion/courseGeneratedFilesQuestion'),
+   ] );
+
+// legacy client file paths
+app.use('/pl/course/:course_id/question/:question_id/file', [
+   require('./middlewares/selectAndAuthzCourseQuestion'),
+   require('./pages/legacyQuestionFile/legacyQuestionFile'),
+   ] );
+app.use('/pl/course/:course_id/question/:question_id/text', [
+   require('./middlewares/selectAndAuthzCourseQuestion'),
+   require('./pages/legacyQuestionText/legacyQuestionText'),
+   ] );
 
 //////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////